+++ src/example.ts
@@ -1,5 +1,5 @@
 import { QvForm, QvInput } from "./validation";
 
-const qvForm = new QvForm("form");
+const qvForm = new QvForm("#myForm");
 
 qvForm.init();